# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_attachment_api_expense_versions__version_id__attachments_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "file",
          "request_id"
        ],
        "title": "Body_upload_attachment_api_expense_versions__version_id__attachments_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PartSearchOut": {
        "additionalProperties": false,
        "properties": {
          "note": {
            "default": "جست‌وجو در ترب باز می‌شود؛ سرور لینک را واکشی نمی‌کند.",
            "title": "Note",
            "type": "string"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "searchUrl": {
            "title": "Searchurl",
            "type": "string"
          }
        },
        "required": [
          "query",
          "searchUrl"
        ],
        "title": "PartSearchOut",
        "type": "object"
      },
...
      "ReceiptDecisionInput": {
        "additionalProperties": false,
        "properties": {
          "approve": {
            "title": "Approve",
            "type": "boolean"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 600,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "approve"
        ],
        "title": "ReceiptDecisionInput",
        "type": "object"
      },
...
      "ReviewExtractionInput": {
        "additionalProperties": false,
        "properties": {
          "lines": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/LineItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lines"
          }
        },
        "title": "ReviewExtractionInput",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "API نمایشی «ترب تعمیر». دادهٔ نمونه، پرداخت شبیه‌سازی‌شده و پاسخ AI نمایشی در این نصب صریحاً برچسب‌گذاری می‌شوند.",
    "title": "ترب تعمیر — API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/attachments/{attachment_id}": {
      "get": {
        "description": "Served by the backend from a private directory — never from a public web path.",
        "operationId": "download_attachment_api_attachments__attachment_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "attachment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Attachment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Attachment",
        "tags": [
          "execution"
        ]
      }
    },
...
    "/api/expense-versions/{version_id}/attachments": {
      "post": {
        "description": "A private evidence image. No OCR, and the image is never sent to the model.",
        "operationId": "upload_attachment_api_expense_versions__version_id__attachments_post",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_attachment_api_expense_versions__version_id__attachments_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Upload Attachment Api Expense Versions  Version Id  Attachments Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Attachment",
        "tags": [
          "execution"
        ]
      }
    },
    "/api/expense-versions/{version_id}/price-checks": {
      "post": {
        "description": "Compare each invoiced part line with the recorded Torob prices.\n\nThis is pure arithmetic over recorded data — it makes no AI call of its own, and on its\nown it changes no agreed amount, no ruling and no refund.",
        "operationId": "run_price_checks_api_expense_versions__version_id__price_checks_post",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PriceCheckOut"
                  },
                  "title": "Response Run Price Checks Api Expense Versions  Version Id  Price Checks Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Price Checks",
        "tags": [
          "execution"
        ]
      }
    },
    "/api/expense-versions/{version_id}/review": {
      "post": {
        "description": "The customer's verdict on this exact version. Rejection opens the dispute path.",
        "operationId": "review_receipt_api_expense_versions__version_id__review_post",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReceiptDecisionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseVersionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review Receipt",
        "tags": [
          "execution"
        ]
      }
    },
    "/api/expense-versions/{version_id}/review-extraction": {
      "post": {
        "description": "The specialist checks the extracted lines before the customer ever sees them.",
        "operationId": "review_extraction_api_expense_versions__version_id__review_extraction_post",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewExtractionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseVersionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review Extraction",
        "tags": [
          "execution"
        ]
      }
    },
...
    "/api/parts/search": {
      "get": {
        "operationId": "part_search_api_parts_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": true,
            "schema": {
              "maxLength": 200,
              "title": "Query",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartSearchOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Part Search",
        "tags": [
          "execution"
        ]
      }
    },
    "/api/parts/{option_id}/confirm-compatibility": {
      "post": {
        "operationId": "confirm_compatibility_api_parts__option_id__confirm_compatibility_post",
        "parameters": [
          {
            "in": "path",
            "name": "option_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Option Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartOptionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Compatibility",
        "tags": [
          "execution"
        ]
      }
    },
...
    "/api/requests/{request_id}/parts": {
      "get": {
        "operationId": "list_parts_api_requests__request_id__parts_get",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PartOptionOut"
                  },
                  "title": "Response List Parts Api Requests  Request Id  Parts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Parts",
        "tags": [
          "execution"
        ]
      },
      "post": {
        "description": "Manually recorded. The server never fetches the link or claims live stock.",
        "operationId": "record_part_api_requests__request_id__parts_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartOptionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartOptionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Part",
        "tags": [
          "execution"
        ]
      }
    },
...
    "/api/requests/{request_id}/price-checks": {
      "get": {
        "operationId": "list_price_checks_api_requests__request_id__price_checks_get",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PriceCheckOut"
                  },
                  "title": "Response List Price Checks Api Requests  Request Id  Price Checks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Price Checks",
        "tags": [
          "execution"
        ]
      }
    },
    "/api/requests/{request_id}/price-snapshots": {
      "get": {
        "operationId": "list_price_snapshots_api_requests__request_id__price_snapshots_get",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PriceSnapshotOut"
                  },
                  "title": "Response List Price Snapshots Api Requests  Request Id  Price Snapshots Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Price Snapshots",
        "tags": [
          "execution"
        ]
      },
      "post": {
        "description": "One recorded seller price. It counts only once support has vouched for it.",
        "operationId": "record_price_snapshot_api_requests__request_id__price_snapshots_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PriceSnapshotInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceSnapshotOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Price Snapshot",
        "tags": [
          "execution"
        ]
      }
    },
...
    "/api/selections/{selection_id}/completion/confirm": {
      "post": {
        "description": "Only when the invoice matches the active agreement and the receipt was confirmed.",
        "operationId": "confirm_completion_api_selections__selection_id__completion_confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "selection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Selection Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmCompletionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompletionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Completion",
        "tags": [
          "execution"
        ]
      }
    },
...
  }
}